Enable HTTP RPC sockets (#871)
* fixed eth start, but crash is not graceful

* better socket connection logic, remove master ps logic for eth

* fix splash screen display of state text for eth

* better gulp download plugin, remove master passwd stuff once and for all

* HTTP socket working

* minor warning

* minor refactor

* had accidentally deleted stuff from fixEthStart merge

* show dialog msg if using http rpc

* update dialog text

* remove git conflict markers

* fix merge
