+++ main.js
@@ -25,6 +25,7 @@ const Settings = require('./modules/settings');
 Settings.init();
 
 
+
 if (Settings.cli.version) {
     console.log(Settings.appVersion);
 
@@ -39,11 +40,18 @@ if (Settings.cli.ignoreGpuBlacklist) {
 // logging setup
 const log = logger.create('main');
 
+
 if (Settings.inAutoTestMode) {
     log.info('AUTOMATED TESTING');
 }
 
 log.info(`Running in production mode: ${Settings.inProductionMode}`);
+
+if ('http' === Settings.rpcMode) {
+    log.warn('Connecting to a node via HTTP instead of IPC. This is less secure!!!!'.toUpperCase());
+}
+
+
 
 
 // db
@@ -160,6 +168,17 @@ var splashWindow;
 // This method will be called when Electron has done everything
 // initialization and ready for creating browser windows.
 app.on('ready', function() {
+    // if using HTTP RPC then inform user
+    if ('http' === Settings.rpcMode) {
+        dialog.showErrorBox('Insecure RPC connection', `
+WARNING: You are connecting to an Ethereum node via: ${Settings.rpcHttpPath}
+
+This is less secure than using local IPC - your passwords will be sent over the wire as plaintext. 
+
+Only do this if you have secured your HTTP connection or you know what you are doing.
+`);
+    }
+
     // initialise the db
     global.db.init().then(onReady).catch((err) => {
         log.error(err);
