+++ modules/ethereumNode.js
@@ -13,7 +13,6 @@ const fs = require('fs');
 const Q = require('bluebird');
 const getNodePath = require('./getNodePath.js');
 const EventEmitter = require('events').EventEmitter;
-const getIpcPath = require('./ipc/getIpcPath.js')
 const Sockets = require('./sockets');
 const Settings = require('./settings');
 
@@ -42,7 +41,7 @@ class EthereumNode extends EventEmitter {
         this._type = null;
         this._network = null;
 
-        this._socket = Sockets.get('node-ipc', Sockets.TYPES.WEB3_IPC);
+        this._socket = Sockets.get('node-ipc', Settings.rpcMode);
 
         this.on('data', _.bind(this._logNodeData, this));
     }
@@ -123,15 +122,7 @@ class EthereumNode extends EventEmitter {
      * @return {Promise}
      */
     init () {
-
-
-        const ipcPath = getIpcPath();
-
-        // TODO: if connection to external node is successful then query it to
-        // determine node and network type
-
-        // check if the node is already running
-        return this._socket.connect({path: ipcPath})
+        return this._socket.connect(Settings.rpcConnectConfig)
             .then(()=> {
                 this.state = STATES.CONNECTED;
 
@@ -262,8 +253,6 @@ class EthereumNode extends EventEmitter {
      * @return {Promise}
      */
     _start (nodeType, network) {
-        const ipcPath = getIpcPath();
-
         log.info(`Start node: ${nodeType} ${network}`);
 
         const isTestNet = ('test' === network);
@@ -295,9 +284,9 @@ class EthereumNode extends EventEmitter {
                 // FORK RELATED
                 this._saveUserData('daoFork', this.daoFork);
 
-                return this._socket.connect({ path: ipcPath }, {
-                        timeout: 30000 /* 30s */
-                    })  
+                return this._socket.connect(Settings.rpcConnectConfig, {
+                    timeout: 30000 /* 30s */
+                })  
                     .then(() => {
                         this.state = STATES.CONNECTED;
                     })
@@ -369,7 +358,7 @@ class EthereumNode extends EventEmitter {
                 // START TESTNET
                 if ('test' == network) {
                     args = (nodeType === 'geth') 
-                        ? ['--testnet', '--fast', '--ipcpath', getIpcPath()] 
+                        ? ['--testnet', '--fast', '--ipcpath', Settings.rpcIpcPath] 
                         : ['--morden', '--unsafe-transactions'];
                 } 
                 // START MAINNET
@@ -527,7 +516,7 @@ class EthereumNode extends EventEmitter {
         try {
             return fs.readFileSync(fullPath, {encoding: 'utf8'});
         } catch (err){
-            log.error(`Unable to read from ${fullPath}`, err);
+            log.warn(`Unable to read from ${fullPath}`, err);
         }
 
         return null;
@@ -542,7 +531,7 @@ class EthereumNode extends EventEmitter {
         try {
             fs.writeFileSync(fullPath, data, {encoding: 'utf8'});
         } catch (err){
-            log.error(`Unable to write to ${fullPath}`, err);
+            log.warn(`Unable to write to ${fullPath}`, err);
         }
     }
